[UPDATE]: Add support for cloning from custom Git providers
This commit adds support for specifying the Git provider (e.g., GitHub or GitLab) when cloning the kernel source. The user can now specify the provider using the `git-provider` input parameter. If no provider is specified, it defaults to GitHub. This enhances the flexibility of the script to work with various Git hosting services.

Signed-off-by: akari <akariondev@gmail.com>